requirements.txt
absl-py==0.9.0
actionlib==1.11.16
adium-theme-ubuntu==0.3.4
angles==1.9.12
astor==0.8.1
attrs==15.2.0
autobahn==0.10.3
backports.weakref==1.0.post1
base-local-planner==1.14.9
beautifulsoup4==4.4.1
bondpy==1.8.3
bzr==2.7.0
camera-calibration==1.12.23
camera-calibration-parsers==1.11.13
catkin==0.7.29
catkin-pkg==0.4.23
catkin-pkg-modules==0.4.23
catkin-tools==0.6.1
certifi==2018.11.29
chardet==3.0.4
configobj==5.0.6
cryptography==1.2.3
cv-bridge==1.12.8
cycler==0.10.0
defusedxml==0.4.1
diagnostic-analysis==1.9.3
diagnostic-common-diagnostics==1.9.3
diagnostic-updater==1.9.3
docutils==0.12
dynamic-reconfigure==1.5.50
easydict==1.6
ecdsa==0.13
empy==3.3.2
enum34==1.1.10
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.3.0
gast==0.3.3
gazebo-plugins==2.5.20
gazebo-ros==2.5.20
gencpp==0.6.0
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.11
gennodejs==2.0.1
genpy==0.6.14
geodesy==0.5.2
glog==0.3.1
gmplot==1.2.0
gps-common==0.3.0
grpcio==1.29.0
h5py==2.8.0
haversine==2.3.0
html5lib==0.999
httplib2==0.9.1
idna==2.8
image-geometry==1.12.8
impacket==0.9.12
interactive-markers==1.11.5
ipaddress==1.0.16
joint-state-publisher==1.12.15
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keyring==7.3
laser-geometry==1.6.5
LatLon==1.0.2
launchpadlib==1.10.3
lazr.restfulclient==0.13.4
lazr.uri==1.0.3
linecache2==1.0.0
lxml==3.5.0
lz4==0.7.0
Markdown==3.1.1
matplotlib==2.0.2
meld==3.14.2
mercurial==3.7.3
message-filters==1.12.17
mock==3.0.5
mpi4py==1.3.1
mpmath==1.0.0
msgpack-python==0.4.6
netifaces==0.10.4
nose==1.3.7
numpy==1.14.5
oauth==1.0.1
opencv-python==3.4.1.15
osrf-pycommon==0.1.9
PAM==0.4.2
pandas==0.24.2
paramiko==1.16.0
pbr==1.8.0
pcapy==0.10.8
Pillow==5.4.1
pip==20.1.1
pluginlib==1.11.3
protobuf==3.12.2
psutil==3.4.2
pyasn1==0.1.9
pyasn1-modules==0.0.7
pycrypto==2.6.1
pydot==1.0.29
pygame==1.9.4
PyGeodesy==19.4.8
Pygments==2.1
pygobject==3.20.0
pygpgme==0.3
PyOpenGL==3.0.2
pyOpenSSL==0.15.1
pyparsing==2.0.3
pyproj==2.1.3
pyserial==3.0.1
Pyste==0.9.10
python-dateutil==2.8.0
python-gflags==3.1.2
python-qt-binding==0.3.7
python-snappy==0.5
pytz==2014.10
pyxdg==0.25
PyYAML==3.11
qt-create==0.2.10
qt-dotgraph==0.3.17
qt-gui==0.3.17
qt-gui-cpp==0.3.17
qt-gui-py-common==0.3.17
requests==2.21.0
resource-retriever==1.12.6
roman==2.0.0
rosbag==1.12.17
rosboost-cfg==1.14.6
rosclean==1.14.6
roscreate==1.14.6
rosdep==0.20.0
rosdep-modules==0.20.0
rosdistro==0.8.3
rosdistro-modules==0.8.3
rosgraph==1.12.17
rosinstall==0.7.8
rosinstall-generator==0.1.22
roslaunch==1.12.17
roslib==1.14.6
roslint==0.11.0
roslz4==1.12.17
rosmake==1.14.6
rosmaster==1.12.17
rosmsg==1.12.17
rosnode==1.12.17
rosparam==1.12.17
rospkg==1.2.9
rospkg-modules==1.2.9
rospy==1.12.17
rosservice==1.12.17
rostest==1.12.17
rostopic==1.12.17
rosunit==1.14.6
roswtf==1.12.17
rqt-action==0.4.9
rqt-bag==0.5.0
rqt-bag-plugins==0.5.0
rqt-console==0.4.9
rqt-dep==0.4.9
rqt-graph==0.4.11
rqt-gui==0.5.0
rqt-gui-py==0.5.0
rqt-image-view==0.4.16
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-multiplot==0.0.10
rqt-nav-view==0.5.7
rqt-plot==0.4.8
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.0
rqt-py-console==0.4.8
rqt-reconfigure==0.5.3
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.13
rqt-robot-steering==0.5.9
rqt-runtime-monitor==0.5.7
rqt-rviz==0.5.10
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.11
rqt-web==0.4.8
rviz==1.12.17
scikit-learn==0.19.1
scipy==1.2.3
SecretStorage==2.1.3
sensor-msgs==1.12.7
service-identity==16.0.0
setuptools==39.1.0
simplejson==3.8.1
six==1.15.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.3
subprocess32==3.5.4
sympy==1.3
tensorboard==1.10.0
tensorflow==1.10.0
tensorflow-gpu==1.10.0
termcolor==1.1.0
testresources==1.0.0
tf==1.11.9
tf-conversions==1.11.9
tf2-geometry-msgs==0.5.20
tf2-kdl==0.5.20
tf2-py==0.5.20
tf2-ros==0.5.20
topic-tools==1.12.17
torch==1.0.1.post2
torchvision==0.2.1
traceback2==1.4.0
trollius==2.0.1
Twisted==16.0.0
txaio==1.0.0
unique-id==1.0.5
unittest2==1.1.0
unity-lens-photos==1.0
urllib3==1.24.1
vcstools==0.1.42
virtualenv==15.0.1
wadllib==1.3.2
Werkzeug==1.0.1
wheel==0.34.2
wstool==0.1.17
wxPython==3.0.2.0
wxPython-common==3.0.2.0
xacro==1.11.3
xlrd==1.1.0
zope.interface==4.1.3
